@@ -2,5 +2,6 @@ package jpademo.dao;
 
 import jpademo.model.Device;
 
-public interface DeviceDao extends CrudDao<Device, Long>{
+public interface DeviceDao extends CrudDao<Device, Long> {
+
 }
