@@ -3,6 +3,12 @@ package jpademo.dao;
 import jpademo.model.Patient;
 import jpademo.model.Test;
 
+import java.time.LocalDate;
+import java.util.List;
+
 public interface TestDao extends CrudDao<Test, Long> {
-    void doSkinTest(Patient patient);
+
+    List<Test> getAllTestsInDateRange(LocalDate from, LocalDate to);
+
+    List<Patient> getPatientsWithGlucoseHigherThan(Double normalLevel);
 }
