@@ -1,6 +1,7 @@
 package jpademo.model;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -11,8 +12,8 @@ import javax.persistence.Table;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Entity
 @Table(name = "photometr_device")
+@Entity
 public class Photometer extends Device {
 
     @Column(name = "type_of_liquid")
