@@ -11,8 +11,8 @@ import javax.persistence.Table;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Entity
 @Table(name = "ultrasonic_accessory")
+@Entity
 public class UltrasonicAccessory extends Accessory {
 
     @Column(name = "radius")
