@@ -1,9 +1,10 @@
 package jpademo.service.impl;
 
 import jpademo.dao.DeviceDao;
-import jpademo.model.Device;
+import jpademo.model.*;
 import jpademo.service.DeviceService;
 
+import java.time.LocalDate;
 import java.util.List;
 
 public class DeviceServiceImpl implements DeviceService {
@@ -38,4 +39,12 @@ public class DeviceServiceImpl implements DeviceService {
     public void delete(Device device) {
         deviceDao.delete(device);
     }
+
+    @Override
+    public void doTest(Patient patient, Test test) {
+        test.setTestDate(LocalDate.now());
+        insert(test.getDevice());
+        patient.doTest(test);
+    }
+
 }
