@@ -5,7 +5,7 @@ import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 
 public class HibernateUtil {
-    private static final String PERSISTENT_UNIT_NAME = "persistence";
+    private static final String PERSISTENT_UNIT_NAME = "persistenceU";
 
     private static final EntityManagerFactory entityManager;
 
